Major updates after laptop installation
Add Zsh, i3, and Polybar configuration for new machine (ash). Reduce the
number of required font packages; most icons now originate from the
material design icon set, some still from Font Awesome.

Replace feh with hsetroot, refer to a symlink in the ~/pictures
directory for the current wallpaper.

Do not run `blueman-applet` and `compton` from the X init script in case
it is not installed.

Make a beginning with scripts that will bootstrap a newly installed Arch
environment.

Remove unused wallpaper.

+++ scripts/.local/bin/run-blueman-applet.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
-blueman-applet &
-while ! pgrep -x blueman-tray; do
-    sleep 0.01
-done
-killall -e blueman-tray
+if [ -x "$(command -v blueman-applet)" ]; then
+    blueman-applet &
+    while ! pgrep -x blueman-tray; do
+        sleep 0.01
+    done
+    killall -e blueman-tray
+fi
